+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 - define a different registry per container
 - config option
+- define commands to excute inside a container
 
 ### Changed